Añadir implementación de la interfaz IRealizacionPedidos

+++ practica03/SupermercadoCommon/src/main/java/es/unican/ps/supermercado/negocio/dominio/LineaPedido.java
@@ -3,19 +3,19 @@ package es.unican.ps.supermercado.negocio.dominio;
 
 public class LineaPedido {
 
-	private double cantidad;
+	private int cantidad;
 	private Articulo articulo;
 	
-	public LineaPedido(double cantidad, Articulo articulo) {
+	public LineaPedido(int cantidad, Articulo articulo) {
 		this.cantidad = cantidad;
 		this.setArticulo(articulo);
 	}
 	
-	public double getCantidad() {
+	public int getCantidad() {
 		return cantidad;
 	}
 	
-	public void setCantidad(double cantidad) {
+	public void setCantidad(int cantidad) {
 		this.cantidad = cantidad;
 	}
